pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bleach>=6.3.0",
    "django>=5.2.8",
    "django-debug-toolbar>=6.1.0",
    "django-filter>=1.1.0",
    "django-rest-framework>=0.1.0",
    "djangorestframework-csv>=3.0.2",
    "djangorestframework-filters>=0.11.1",
    "djangorestframework-xml>=2.0.0",
    "djangorestframework-yaml>=2.0.0",
]
